+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.12.0] - 2017-12-18
+### Added
+ - ``$upper`` feature for SELECT Statements
+ - ``$lower`` feature for SELECT Statements
+ - ``$function`` feature for SELECT Statements
 
 ## [0.11.1] - 2017-03-20
 ### Added